review: dry up some common code
